Moved Die and RandomDie to a shared location, so it's more natural that both loot and combat are using them
