+++ src/main/java/com/projectswg/holocore/resources/support/random/Die.kt
@@ -24,10 +24,16 @@
  * You should have received a copy of the GNU Affero General Public License        *
  * along with Holocore.  If not, see <http://www.gnu.org/licenses/>.               *
  ***********************************************************************************/
-package com.projectswg.holocore.services.gameplay.combat.command
+package com.projectswg.holocore.resources.support.random
 
-class RandomDie : Die {
-	override fun roll(range: IntRange): Int {
-		return range.random()
-	}
+/**
+ * A die is a random number generator that can be used to simulate dice rolls.
+ * The abstraction exists to allow for testing, where the die can be replaced with a deterministic implementation.
+ */
+interface Die {
+
+	/**
+	 * Rolls the die in the specified range.
+	 */
+	fun roll(range: IntRange): Int
 }
